requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
motor==3.1.2
python-jose==3.3.0
passlib==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.24.1
jinja2==3.1.2
bson==0.5.10
pymongo==4.3.3
